+++ dumpling/cli/versions.go
@@ -18,7 +18,7 @@ import (
 
 	"go.uber.org/zap"
 
-	"github.com/pingcap/tidb/dumpling/v4/log"
+	"github.com/pingcap/tidb/dumpling/log"
 )
 
 var (
