+++ dumpling/export/block_allow_list.go
@@ -5,7 +5,7 @@ package export
 import (
 	"go.uber.org/zap"
 
-	tcontext "github.com/pingcap/tidb/dumpling/v4/context"
+	tcontext "github.com/pingcap/tidb/dumpling/context"
 )
 
 func filterDatabases(tctx *tcontext.Context, conf *Config, databases []string) []string {
