+++ dumpling/export/config_test.go
@@ -5,7 +5,7 @@ package export
 import (
 	"testing"
 
-	tcontext "github.com/pingcap/tidb/dumpling/v4/context"
+	tcontext "github.com/pingcap/tidb/dumpling/context"
 	"github.com/stretchr/testify/require"
 )
