+++ dumpling/export/dump.go
@@ -21,9 +21,9 @@ import (
 	pclog "github.com/pingcap/log"
 	"github.com/pingcap/tidb/br/pkg/storage"
 	"github.com/pingcap/tidb/br/pkg/summary"
-	"github.com/pingcap/tidb/dumpling/v4/cli"
-	tcontext "github.com/pingcap/tidb/dumpling/v4/context"
-	"github.com/pingcap/tidb/dumpling/v4/log"
+	"github.com/pingcap/tidb/dumpling/cli"
+	tcontext "github.com/pingcap/tidb/dumpling/context"
+	"github.com/pingcap/tidb/dumpling/log"
 	"github.com/pingcap/tidb/store/helper"
 	"github.com/pingcap/tidb/tablecodec"
 	"github.com/pingcap/tidb/util/codec"
