+++ dumpling/export/http_handler.go
@@ -10,8 +10,8 @@ import (
 	"time"
 
 	"github.com/pingcap/errors"
-	tcontext "github.com/pingcap/tidb/dumpling/v4/context"
-	"github.com/pingcap/tidb/dumpling/v4/log"
+	tcontext "github.com/pingcap/tidb/dumpling/context"
+	"github.com/pingcap/tidb/dumpling/log"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/soheilhy/cmux"
 )
