+++ dumpling/export/ir.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 
 	"github.com/pingcap/errors"
-	tcontext "github.com/pingcap/tidb/dumpling/v4/context"
+	tcontext "github.com/pingcap/tidb/dumpling/context"
 )
 
 // TableDataIR is table data intermediate representation.
