+++ dumpling/export/prepare.go
@@ -10,7 +10,7 @@ import (
 	"text/template"
 
 	"github.com/pingcap/errors"
-	tcontext "github.com/pingcap/tidb/dumpling/v4/context"
+	tcontext "github.com/pingcap/tidb/dumpling/context"
 )
 
 const (
