+++ dumpling/export/prepare_test.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"testing"
 
-	tcontext "github.com/pingcap/tidb/dumpling/v4/context"
+	tcontext "github.com/pingcap/tidb/dumpling/context"
 	"github.com/stretchr/testify/require"
 
 	"github.com/DATA-DOG/go-sqlmock"
