+++ dumpling/export/util.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/pingcap/errors"
-	tcontext "github.com/pingcap/tidb/dumpling/v4/context"
+	tcontext "github.com/pingcap/tidb/dumpling/context"
 	"go.etcd.io/etcd/clientv3"
 )
